rename filenames from ClientSession to Client (not classname)
