+++ Makefile
@@ -1,6 +1,6 @@
 NAME := ircserv
 
-SRCS := src/main.cpp src/IRCServer.cpp src/ClientSession.cpp \
+SRCS := src/main.cpp src/IRCServer.cpp src/Client.cpp \
 		src/CommandHandler.cpp src/IRCMessage.cpp src/Socket.cpp \
 		src/IRCParser.cpp src/Utils.cpp src/IOWrapper.cpp src/IRCLogger.cpp
 OBJS := $(SRCS:.cpp=.o)
